+++ src/main/java/online/kingdomkeys/kingdomkeys/block/ModBlocks.java
@@ -40,6 +40,7 @@ public class ModBlocks {
             pairBlox = createNewBlock("pair_blox", () -> new PairBloxBlock(Block.Properties.of().mapColor(MapColor.COLOR_LIGHT_GREEN).strength(1.0F, 10.0F))),
             infestedNormalBlox = createNewBlock("infested_normal_blox", () -> new InfestedNormalBlox(Block.Properties.of().mapColor(MapColor.COLOR_RED).strength(1, 10))),
             magicTarget = createNewBlock("magic_target", () -> new MagicTargetBlock(Block.Properties.of().mapColor(MapColor.COLOR_RED).strength(1, 10))),
+            flowmotionRail = createNewBlock("flowmotion_rail", () -> new FlowmotionRailBlock(Block.Properties.of().mapColor(MapColor.METAL).noCollission().strength(0.7F).sound(SoundType.METAL))),
 
             //TODO HARVEST LEVEL REPLACED BY TAGS
             blazingOre = createNewBlock("blazing_ore", () -> new KKOreBlock(Block.Properties.of().mapColor(MapColor.STONE).instrument(NoteBlockInstrument.BASEDRUM).strength(2.0F, 1.0F).lightLevel((state) -> 4))), //HL 1