+++ src/main/java/online/kingdomkeys/kingdomkeys/datagen/init/LanguageENUS.java
@@ -586,6 +586,7 @@ public class LanguageENUS extends KKLanguageProvider {
         addBlock(ModBlocks.prizeBlox, "Prize Blox");
         addBlock(ModBlocks.rarePrizeBlox, "Rare Prize Blox");
         addBlock(ModBlocks.magnetBlox, "Magnet Blox");
+        addBlock(ModBlocks.flowmotionRail, "Flowmotion Rail");
         addBlock(ModBlocks.pairBlox, "Pair Blox");
         addBlock(ModBlocks.infestedNormalBlox, "Infested Normal Blox");
         addBlock(ModBlocks.gummiMeteor, "Gummi Meteorite");