+++ src/main/java/online/kingdomkeys/kingdomkeys/datagen/init/LanguageESES.java
@@ -845,6 +845,7 @@ public class LanguageESES extends KKLanguageProvider {
         addBlock(ModBlocks.prizeBlox, "Bloque de Premio");
         addBlock(ModBlocks.rarePrizeBlox, "Bloque de Premio Raro");
         addBlock(ModBlocks.magnetBlox, "Bloque Magn\u00e9tico");
+        addBlock(ModBlocks.flowmotionRail, "Raíl de Flowmotion");
         addBlock(ModBlocks.pairBlox, "Bloque Emparejado");
         addBlock(ModBlocks.infestedNormalBlox, "Bloque Normal Infestado");
         addBlock(ModBlocks.gummiMeteor, "Meteorito Gummi");